feat(logout):add logout feature in side drawer -UL-181 - add logout functions in SideDrawer - refactor navigator to named routes to avoid pages routes misplaced
